backend/api/requirements.txt
fastapi
uvicorn
google-generativeai
joblib
requests
python-dotenv